+++ src/MVHub/main.cpp
@@ -1,10 +1,12 @@
 #include "MVHub.h"
 #include "Application.h"
+#include <QStyleFactory>
 
 int main(int argc, char *argv[])
 {
 	Application a(argc, argv);
 	MVHub w;
+	w.setStyle(QStyleFactory::create("windowsvista"));
 	w.showMaximized();
 	
 	return a.exec();