+++ src/cppNGS/VariantHgvsAnnotator.h
@@ -59,6 +59,8 @@ struct CPPNGSSHARED_EXPORT VariantConsequence
 	QByteArray normalized; //normalized VCF representation after shifting according to 3' rule
 
 	QByteArray typesToString(QByteArray sep="&") const;
+	//Only shows most important type
+	QByteArray typesToStringSimplified(QByteArray sep="&") const;
 	static QByteArray typeToString(VariantConsequenceType type)
     {
         switch(type)
@@ -93,6 +95,9 @@ struct CPPNGSSHARED_EXPORT VariantConsequence
 
 		THROW(ProgrammingException, "Unhandled variant consequence type " + QByteArray::number(static_cast<int>(type)) + "!");
     }
+
+	//Returns exon/intron number an total number of exons/introns
+	QByteArray exonOrIntron(const Transcript& trans);
 
 	QByteArray toString() const;
 };