+++ src/cppNGSD/SomaticcfDNAReport.cpp
@@ -139,7 +139,7 @@ RtfTable SomaticcfDnaReport::partSnvTable(int cfdna_idx_start, int cfdna_idx_end
 		if (co_sp_anno.trans.isValid())
 		{
 			change_string = co_sp_anno.consequence.hgvs_c + ", " + co_sp_anno.consequence.hgvs_p + "\n\\line\n" + co_sp_anno.trans.nameWithVersion();
-			type_string = co_sp_anno.consequence.typesToString();
+			type_string = co_sp_anno.consequence.typesToStringSimplified();
 		}
 		else
 		{